CharArray should be consistent with mt5

+++ src/client_api/fx_action.h
@@ -1,9 +1,9 @@
 #ifndef _CLIENT_API_FX_ACTION_H_
 #define _CLIENT_API_FX_ACTION_H_
-#include "messenger/msg.h"
+#include "win_messenger/msg.h"
 
 enum class FXAct {
-    HISTORY = MsgAct::NUM_ACTIONS,
+    HISTORY = (int)MsgAct::NUM_ACTIONS,
     HISTORY_MINBAR,
     //CHECKIN,
     TICK,
